Fixed issue #6 (not optimal)
The layouts for the text blocks are not cleaned up when the blocks are
removed from the document because they are linked to the console display
map. This fix is not the fastest way to execute the display code, but it
does still allow the layouts to be used in hit tests.